requirements.txt
datasets>=2.18.0
PyYAML>=6.0.1
pytest>=8.2.0
dspy-ai>=2.5.25
nbformat>=5.10.4
nbconvert>=7.16.4
jupyter>=1.1.0
